+++ weblate/accounts/tests/test_views.py
@@ -110,7 +110,7 @@ class ViewTest(TestCase):
         self.assertContains(response, 'Enter a valid email address.')
 
     @override_settings(
-        AUTH_MAX_ATTEMPTS=0,
+        RATELIMIT_ATTEMPTS=0,
     )
     def test_contact_rate(self):
         """Test for contact form rate limiting."""
@@ -261,7 +261,7 @@ class ViewTest(TestCase):
             'This username/password combination was not found.'
         )
 
-    @override_settings(AUTH_MAX_ATTEMPTS=20, AUTH_LOCK_ATTEMPTS=5)
+    @override_settings(RATELIMIT_ATTEMPTS=20, AUTH_LOCK_ATTEMPTS=5)
     def test_login_ratelimit(self, login=False):
         if login:
             self.test_login()
@@ -283,7 +283,7 @@ class ViewTest(TestCase):
         )
         self.assertContains(response, 'Please try again.')
 
-    @override_settings(AUTH_MAX_ATTEMPTS=10, AUTH_LOCK_ATTEMPTS=5)
+    @override_settings(RATELIMIT_ATTEMPTS=10, AUTH_LOCK_ATTEMPTS=5)
     def test_login_ratelimit_login(self):
         self.test_login_ratelimit(True)