+++ docs/changes.rst
@@ -7,6 +7,7 @@ weblate 3.2
 Released on ? 2018.
 
 * Add install_addon management command for automated addon installation.
+* Allow more fine grained ratelimit settings.
 
 weblate 3.1.1
 -------------